pyproject.toml
[tool.poetry]
name = "synthetic-ap"
version = "0.1.0"
description = "Stage-1 unpaid bills generator & Xero inserter"
authors = ["RedOwl AI"]
packages = [{ include = "synthap", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
plotly = "^5.18.0"
pydantic = "^2.7"
pydantic-settings = "^2.4"
pyyaml = "^6.0.2"
typer = {version="==0.12.3", extras=["all"]}
click = "==8.1.7"
python-dateutil = "^2.9"
pandas = "^2.2"
duckdb = "^1.0.0"
pyarrow = "^17.0.0"
httpx = "^0.27.0"
tenacity = "^8.5.0"
tqdm = "^4.66.4"
numpy = "^2.0.1"
tzdata = "^2024.1"
python-slugify = "^8.0.4"
orjson = "^3.10.7"
openai = "^1.40.0"
fastapi = "^0.114.0"
uvicorn = {version="^0.30.6", extras=["standard"]}
python-dotenv = "^1.1.1"
streamlit = "^1.37.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
black = "^24.8.0"
ruff = "^0.6.3"
mypy = "^1.11.1"

[tool.ruff]
line-length = 100
select = ["E","F","I","UP","B"]
fix = true
